deepen: inline openrouter-credit-monitor into daily-report-handler
`lib/ai/openrouter-credit-monitor.ts` was a 160-LOC module with two
exports (`getOpenRouterCreditStatus`, `formatCreditStatusForSlack`)
and one shared type (`CreditStatus`). One production caller:
`lib/slack/daily-report-handler.ts`. No dedicated test file — the
module was never exercised through its interface in isolation.

The interface was nearly as complex as the implementation. The fetch
returns a shape (`CreditStatus`) whose only downstream consumer is
the formatter in the same file, and the formatter's return shape
(`{text, emoji, color}`) is dictated entirely by the Slack block the
caller then emits. Neither seam varies: one adapter, one caller —
per LANGUAGE.md, a hypothetical seam, not a real one.

Fold both exports and the private types/threshold constant into
`daily-report-handler.ts` as module-private members alongside the
other section-metric helpers the file already carries. The daily-
report module keeps its existing interface (`handleDailyReport`,
`handleIntervalReport`, etc.); credit-status is now one of its
internal signals rather than a cross-module import.

No test rewiring needed: no test file mocked the module. The single
existing daily-report test (`interval-report.test.ts`) still passes.
